⚡️ v2.7.8 Memory improvements, add precision and scale support to numerics. (#197)
# v2.7.8

- **Add support for user-supplied precision and scale for `numeric` columns.**  
  You may now manually specify a numeric column's precision and scale:

  ```python
  import meerschaum as mrsm

  pipe = mrsm.Pipe(
      'demo', 'numeric', 'precision_scale',
      instance='sql:local',
      dtypes={'val': 'numeric[5,2]'},
  )
  pipe.sync([{'val': '123.456'}])
  print(pipe.get_data())
  #       val
  # 0  123.46
  ```

- **Serialize `numeric` columns to exact values during bulk inserts.**  
  Decimal values are serialized when inserting into `NUMERIC` columns during bulk inserts.

- **Return a generator when fetching with `SQLConnector`.**  
  To alleviate memory pressure, skip loading the entire dataframe when fetching.

- **Add `json_serialize_value()` to handle custom dtypes.**  
  When serializing documents, pass `json_serialize_value` as the default handler:

  ```python
  import json
  from decimal import Decimal
  from datetime import datetime, timezone
  from meerschaum.utils.dtypes import json_serialize_value

  print(json.dumps(
      {
          'bytes': b'hello, world!',
          'decimal': Decimal('1.000000001'),
          'datetime': datetime(2025, 1, 1, tzinfo=timezone.utc),
      },
      default=json_serialize_value,
      indent=4,
  ))
  # {
  #     "bytes": "aGVsbG8sIHdvcmxkIQ==",
  #     "decimal": "1.000000001",
  #     "datetime": "2025-01-01T00:00:00+00:00"
  # }
  ```

- **Fix an issue with the `WITH` keyword in pipe definitions for MSSQL.**  
  Previously, pipes with used with keyword `WITH` but not as a CTE (e.g. to specify an index) were incorrectly parsed.

+++ meerschaum/utils/daemon/Daemon.py
@@ -774,9 +774,16 @@ class Daemon:
         if '_process' not in self.__dict__ or self.__dict__['_process'].pid != int(pid):
             try:
                 self._process = psutil.Process(int(pid))
+                process_exists = True
             except Exception:
-                if self.pid_path.exists():
-                    self.pid_path.unlink()
+                process_exists = False
+            if not process_exists:
+                _ = self.__dict__.pop('_process', None)
+                try:
+                    if self.pid_path.exists():
+                        self.pid_path.unlink()
+                except Exception:
+                    pass
                 return None
         return self._process
